requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
cPython==0.0.6
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
docker-pycreds==0.4.0
fonttools==4.28.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
seaborn==0.11.2
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thop==0.0.31.post2005241907
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
wandb==0.12.7
Werkzeug==2.0.2
yaspin==2.1.0
zipp==3.6.0

# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# Logging -------------------------------------
# tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
# thop  # FLOPs computation
